requirements.txt
# Core
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0

# Vector store
chromadb>=0.5.0

# S3 / object storage
boto3>=1.34.0

# Config
python-dotenv>=1.0.0
